pyproject.toml
[tool.poetry]
name = "e-commerce-mvp"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "e_commerce_mvp"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.1"
uvicorn = "^0.24.0.post1"
tortoise-orm = {extras = ["asyncpg"], version = "^0.20.0"}
aerich = "^0.7.2"
passlib = "^1.7.4"
pydantic = {extras = ["email"], version = "^2.4.2"}
bcrypt = "^4.0.1"
python-jose = "^3.3.0"
pyjwt = "^2.8.0"
python-multipart = "^0.0.6"
python-dotenv = "^1.0.0"
flake8 = "^6.1.0"
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
httpx = "^0.25.1"
coverage = "^7.3.2"



[tool.aerich]
tortoise_orm = "app.core.config.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
